Tested import service against local geoserver and everything is working. Beefed up error checking a bit more. Still needs loads of error checking

+++ coastal-hazards-ui/src/main/java/gov/usgs/cida/coastalhazards/service/ImportService.java
@@ -5,22 +5,28 @@ import gov.usgs.cida.utilities.communication.GeoserverHandler;
 import gov.usgs.cida.utilities.communication.RequestResponseHelper;
 import gov.usgs.cida.utilities.properties.JNDISingleton;
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.StringReader;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import org.apache.commons.codec.binary.Base64OutputStream;
+import javax.xml.xpath.XPath;
+import javax.xml.xpath.XPathConstants;
+import javax.xml.xpath.XPathExpressionException;
+import javax.xml.xpath.XPathFactory;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.io.IOUtils;
-import org.apache.commons.lang.StringEscapeUtils;
 import org.apache.commons.lang.StringUtils;
+import org.apache.http.HttpResponse;
 import org.slf4j.LoggerFactory;
+import org.w3c.dom.NodeList;
+import org.xml.sax.InputSource;
 
 /**
  *
@@ -41,7 +47,7 @@ public class ImportService extends HttpServlet {
     private static String GEOSERVER_USER_PARAM_CONFIG_KEY = "coastal-hazards.geoserver.username";
     private static String GEOSERVER_PASS_PARAM_CONFIG_KEY = "coastal-hazards.geoserver.password";
     private static GeoserverHandler geoserverHandler = null;
-    
+
     @Override
     public void init() throws ServletException {
         super.init();
@@ -56,71 +62,71 @@ public class ImportService extends HttpServlet {
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         String fileToken = request.getParameter("file-token");
+        String featureName = request.getParameter("feature-name");
+        String workspace = request.getParameter("workspace");
         File fileDirectoryHandle = new File(new File(uploadDirectory), fileToken);
         File shapeFile;
+        String name;
         Map<String, String> responseMap = new HashMap<String, String>();
-        
+
         if (StringUtils.isBlank(fileToken)) {
             responseMap.put("error", "Parameter 'file-token' cannot be blank");
             RequestResponseHelper.sendErrorResponse(response, responseMap);
             return;
         }
-        
+
+        if (StringUtils.isBlank(workspace)) {
+            responseMap.put("error", "Parameter 'workspace' cannot be blank");
+            RequestResponseHelper.sendErrorResponse(response, responseMap);
+            return;
+        }
+
         if (!fileDirectoryHandle.exists() || fileDirectoryHandle.listFiles().length == 0) {
             FileUtils.deleteQuietly(fileDirectoryHandle);
             responseMap.put("error", "The file denoted by token " + fileToken + " does not exist.");
             RequestResponseHelper.sendErrorResponse(response, responseMap);
             return;
         }
-        
+
         // We assume there is only one file per directory
         shapeFile = fileDirectoryHandle.listFiles()[0];
-        String importResult = geoserverHandler.importFeatures(shapeFile, "ch-input", FilenameUtils.removeExtension(shapeFile.getName()));
+        name = StringUtils.isBlank(featureName) ? FilenameUtils.removeExtension(shapeFile.getName()) : featureName;
+        HttpResponse importResponse = geoserverHandler.importFeatures(shapeFile, workspace, name);
+
+        String responseText = IOUtils.toString(importResponse.getEntity().getContent());
         
-        responseMap.put("file-token", fileToken);
-        responseMap.put("endpoint", importResult);
-        RequestResponseHelper.sendSuccessResponse(response, responseMap);
-        
+        if (!responseText.toLowerCase().contains("ows:exception")) {
+            responseMap.put("file-token", fileToken);
+            responseMap.put("feature", responseText);
+            responseMap.put("endpoint", geoserverEndpoint);
+            RequestResponseHelper.sendSuccessResponse(response, responseMap);
+        } else {
+            InputSource source = new InputSource(new StringReader(responseText));
+            XPath xPath = XPathFactory.newInstance().newXPath();
+            NodeList list;
+            String error = "";
+            try {
+                list = (NodeList) xPath.evaluate("//*[local-name()='ExceptionText']", source, XPathConstants.NODESET);
+                error = list.item(0).getTextContent();
+            } catch (XPathExpressionException ex) {
+                Logger.getLogger(ImportService.class.getName()).log(Level.SEVERE, null, ex);
+            }
+
+            responseMap.put("file-token", fileToken);
+            responseMap.put("error", error);
+            RequestResponseHelper.sendErrorResponse(response, responseMap);
+        }
     }
 
-    // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
-    /**
-     * Handles the HTTP
-     * <code>GET</code> method.
-     *
-     * @param request servlet request
-     * @param response servlet response
-     * @throws ServletException if a servlet-specific error occurs
-     * @throws IOException if an I/O error occurs
-     */
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         processRequest(request, response);
     }
 
-    /**
-     * Handles the HTTP
-     * <code>POST</code> method.
-     *
-     * @param request servlet request
-     * @param response servlet response
-     * @throws ServletException if a servlet-specific error occurs
-     * @throws IOException if an I/O error occurs
-     */
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         processRequest(request, response);
     }
-
-    /**
-     * Returns a short description of the servlet.
-     *
-     * @return a String containing servlet description
-     */
-    @Override
-    public String getServletInfo() {
-        return "Short description";
-    }// </editor-fold>
 }
